backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jalexanderII", email = "a.joeljr@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "cachetools>=6.2.0",
    "fastapi>=0.116.1",
    "httpx[http2]>=0.28.1",
    "polygon-api-client>=1.15.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
